Remove old components and constrain the styled ones
