--- src/components/atoms/Button/index.ts
@@ -1,5 +1,0 @@
-/**
- * @deprecated Use named export instead of default export
- */
-export { default } from "./Button";
-export { AppButton as Button } from "./Button";
