+++ src/components/atoms/ButtonOG/index.ts
@@ -1,0 +1,1 @@
+export { ButtonOG } from "./ButtonOG";
