--- src/components/atoms/Legend/Legend.types.ts
@@ -1,6 +1,0 @@
-export interface LegendProps {
-  text: string;
-  position?: "left" | "right";
-  onClick?: () => void;
-  icon?: Element;
-}
