--- src/components/molecules/TabNavigation/TabNavigation.scss
@@ -1,25 +1,0 @@
-@import "scss/constants.scss";
-
-.tab-navigation-container {
-  border-radius: 22px;
-  background: $secondary--live;
-  display: flex;
-  flex-shrink: 0;
-  margin-bottom: 10px;
-  width: 90%;
-
-  .tab-item {
-    display: flex;
-    justify-content: center;
-    border-radius: 22px;
-    font-weight: 700;
-    font-size: 16px;
-    padding: 13px 10px;
-    cursor: pointer;
-  }
-
-  .selected-tab-item {
-    background: $primary--over;
-    color: $content--under;
-  }
-}
