--- src/components/molecules/TabNavigation/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./TabNavigation";
