--- src/components/molecules/TableFooter/index.ts
@@ -1,1 +1,0 @@
-export { TableFooter } from "./TableFooter";
