+++ src/components/molecules/TableHeader/index.ts
@@ -1,6 +1,1 @@
 export { TableHeader } from "./TableHeader";
-
-/**
- * @deprecated use named export instead
- */
-export { default } from "./TableHeader";
