--- src/components/organisms/BannerAdmin/BannerAdmin.scss
@@ -1,51 +1,0 @@
-@import "scss/constants.scss";
-
-.BannerAdmin {
-  width: 100%;
-
-  &__input-text {
-    text-align: start;
-    padding: $spacing--sm $spacing--lg;
-    font-size: $font-size--md;
-
-    &--large {
-      height: $announcement-banner-admin-height;
-    }
-  }
-
-  &__input-container {
-    margin-bottom: $spacing--lg;
-  }
-
-  &__button-container {
-    display: flex;
-    justify-content: space-between;
-    margin-top: $spacing--xl;
-  }
-
-  &__button {
-    width: $announce-button-width;
-  }
-
-  &__action-container {
-    margin-left: 50px;
-  }
-
-  &__checkbox {
-    justify-content: start;
-    margin-bottom: $spacing--lg;
-
-    &--action {
-      margin-bottom: $spacing--xs;
-    }
-  }
-
-  &__checkbox__label {
-    font-size: $font-size--lg;
-  }
-
-  &__checkbox__label__disabled {
-    color: --white;
-    opacity: 20%;
-  }
-}
