--- src/components/organisms/BannerAdmin/index.ts
@@ -1,1 +1,0 @@
-export { BannerAdmin } from "./BannerAdmin";
