--- src/components/organisms/DustStorm/DustStorm.scss
@@ -1,102 +1,0 @@
-@import "scss/constants";
-
-$side-padding: 30px;
-
-$border-radius: 28px;
-$large-shadow: 0 10px 30px 0 rgba(0, 0, 0, 0.34);
-$modal-max-width: 540px;
-
-.duststorm-container {
-  position: absolute;
-  z-index: z(duststorm-container);
-  display: flex;
-  align-items: center;
-  overflow: hidden;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  /* Rectangle 18: */
-  background: rgba(253, 166, 130, 0.56);
-  backdrop-filter: blur(2px);
-  display: flex;
-  opacity: 0;
-  transition: all 700ms ease-in;
-
-  &.show {
-    opacity: 1;
-  }
-  &:after {
-    animation: dust 8s steps(10) infinite;
-    background-image: url(./dust-pattern.png);
-    content: "";
-    height: 300%;
-    left: -50%;
-    opacity: 0.1;
-    position: fixed;
-    top: -110%;
-    width: 300%;
-  }
-
-  @keyframes dust {
-    0%,
-    100% {
-      transform: translate(0, 0);
-    }
-    10% {
-      transform: translate(-5%, -10%);
-    }
-    20% {
-      transform: translate(-15%, 5%);
-    }
-    30% {
-      transform: translate(7%, -25%);
-    }
-    40% {
-      transform: translate(-5%, 25%);
-    }
-    50% {
-      transform: translate(-15%, 10%);
-    }
-    60% {
-      transform: translate(15%, 0%);
-    }
-    70% {
-      transform: translate(0%, 15%);
-    }
-    80% {
-      transform: translate(3%, 35%);
-    }
-    90% {
-      transform: translate(-10%, 10%);
-    }
-  }
-
-  .modal-content {
-    z-index: z(duststorm-modal-content);
-    width: calc(100% - 20px);
-    max-width: $modal-max-width;
-    margin: 0 auto;
-    padding: $side-padding;
-    background-color: $content--under;
-    box-shadow: $large-shadow;
-    text-align: center;
-    border-radius: $border-radius;
-    overflow: hidden;
-    opacity: 1;
-    transition: all 600ms $transition-function;
-    transform: translateY(100px);
-    &.show {
-      opacity: 1;
-      transform: translateY(0px);
-    }
-
-    h3 {
-      margin-bottom: 1rem;
-    }
-
-    p {
-      margin-bottom: 1rem;
-    }
-  }
-}
