+++ src/components/templates/FireBarrel/FireBarrel.tsx
@@ -1,4 +1,5 @@
 import React, { useMemo } from "react";
+import styled, { css } from "styled-components";
 
 import { AnyVenue } from "types/venues";
 
@@ -12,7 +13,85 @@ import { VideoParticipant } from "components/organisms/Video";
 
 import { LoadingPage } from "components/molecules/LoadingPage/LoadingPage";
 
-import * as S from "./FireBarrel.styled";
+const StyledWrapper = styled.div`
+  display: grid;
+  width: 60%;
+  margin: 6rem auto;
+  grid-template-columns: repeat(4, 1fr);
+  grid-gap: 1rem;
+`;
+
+const StyledBarrel = styled.iframe.attrs({
+  title: "FireBarrelVideo",
+  frameBorder: "0",
+  allow:
+    "accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture",
+})`
+  width: 100%;
+  height: 100%;
+  grid-column: 2 / span 2;
+  grid-row: 1 / span 2;
+  align-self: center;
+  justify-self: center;
+
+  border-radius: 28px;
+`;
+
+const emptyChair = css`
+  &:after {
+    content: "+";
+    font-size: 2rem;
+  }
+`;
+
+const filledChair = css`
+  background-color: unset;
+`;
+
+const StyledChair = styled.div<{
+  isEmpty?: boolean;
+}>`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+
+  position: relative;
+
+  border-radius: 28px;
+
+  background-color: #1a1d24;
+
+  overflow: hidden;
+
+  ${({ isEmpty }) => (isEmpty ? emptyChair : filledChair)};
+
+  &:before {
+    content: "";
+    display: inline-block;
+    width: 1px;
+    height: 0;
+    padding-bottom: calc(100% / (1 / 1));
+  }
+
+  .col {
+    width: 100%;
+    height: 100%;
+    padding: 0;
+  }
+
+  .av-controls {
+    display: flex;
+    justify-content: space-around;
+    margin: 4px;
+
+    position: absolute;
+    right: 0;
+    bottom: 1em;
+    left: 0;
+
+    text-align: center;
+  }
+`;
 
 const DEFAULT_BURN_BARREL_SEATS = 8;
 
@@ -39,8 +118,8 @@ export const FireBarrel: React.FC<FireBarrelProps> = ({ venue }) => {
   if (!userWithId) return <LoadingPage />;
 
   return (
-    <S.Wrapper>
-      <S.Barrel
+    <StyledWrapper>
+      <StyledBarrel
         src={convertToEmbeddableUrl({
           url: venue?.iframeUrl,
           autoPlay: true,
@@ -51,30 +130,30 @@ export const FireBarrel: React.FC<FireBarrelProps> = ({ venue }) => {
           participants?.[index] ?? {};
 
         if (!participantUserData) {
-          return <S.Chair key={index} isEmpty />;
+          return <StyledChair key={index} isEmpty />;
         }
 
         const isMe = participantUserData.id === userId;
 
         if (!!localParticipant && isMe) {
           return (
-            <S.Chair key={userId}>
+            <StyledChair key={userId}>
               <VideoParticipant
                 participant={localParticipant}
                 participantUser={userWithId}
               />
-            </S.Chair>
+            </StyledChair>
           );
         }
 
         if (participants.length && !!participants[index]) {
           return (
-            <S.Chair key={participant.identity}>
+            <StyledChair key={participant.identity}>
               <VideoParticipant
                 participant={participant}
                 participantUser={userWithId}
               />
-            </S.Chair>
+            </StyledChair>
           );
         }
 
@@ -82,6 +161,6 @@ export const FireBarrel: React.FC<FireBarrelProps> = ({ venue }) => {
       })}
 
       {renderErrorModal()}
-    </S.Wrapper>
+    </StyledWrapper>
   );
 };
